Add files via upload
Introduction
This is a Stock Management System in PHP and MySQL Database. This is a web-based application that provides an online and automated platform for shops or businesses. This project can manage the company's Purchase Orders, Receiving, Back Orders, returns, and Sales Records. The application has a pleasant user interface with the help of Bootstrap Library and AdminLTE template. This has also user-friendly functionalities.